TEC-6428: refactoring packages
