+++ src/test/java/com/echobox/api/linkedin/jsonmapper/DefaultJsonMapperTest.java
@@ -39,8 +39,8 @@ import com.echobox.api.linkedin.types.objectype.MultiLocaleString;
 import com.echobox.api.linkedin.types.social.actions.CommentAction;
 import com.echobox.api.linkedin.types.social.actions.LikeAction;
 import com.echobox.api.linkedin.types.social.actions.SocialAction;
-import com.echobox.api.linkedin.types.statistics.Statistics;
-import com.echobox.api.linkedin.types.statistics.TotalPageStatistics;
+import com.echobox.api.linkedin.types.statistics.page.Statistics;
+import com.echobox.api.linkedin.types.statistics.page.TotalPageStatistics;
 import com.echobox.api.linkedin.types.urn.function.FunctionURN;
 import com.echobox.api.linkedin.types.urn.location.CountryGroupURN;
 import com.echobox.api.linkedin.types.urn.location.CountryURN;
